Added the CreateEncounterAction to allow adding of encounters via incidents
