+++ Assets/Scripts/Incidents/IncidentActions/ActionFields/IncidentActionFieldContainer.cs
@@ -10,13 +10,22 @@ namespace Game.Incidents
 	[HideReferenceObjectPicker]
 	public class IncidentActionFieldContainer
 	{
-		[ValueDropdown("GetFilteredTypeList"), OnValueChanged("SetContextType"), LabelText("Context Type")]
+		[ValueDropdown("GetFilteredTypeList"), OnValueChanged("SetContextType"), LabelText("Context Type"), ShowIf("@this.showTypeDropdown == true")]
 		public Type contextType;
 		[ShowIf("@this.actionField != null")]
 		public IIncidentActionField actionField;
 
 		[HideInInspector, JsonIgnore]
 		public Action onSetContextType;
+
+		private bool showTypeDropdown = true;
+
+		public void ForceSetContextType(Type type)
+		{
+			contextType = type;
+			showTypeDropdown = false;
+			SetContextType();
+		}
 
 		private void SetContextType()
 		{
