+++ Assets/Scripts/Incidents/IncidentActions/Actions/CreateEncounterAction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: acca2da93a9507449b54f1ca8a819c70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
