+++ Assets/Scripts/Simulation/Adventures/AdventureEncounter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 084bf99f985f0d341a9f6e9f9a93e27c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
