+++ Assets/Scripts/Simulation/Adventures/AdventureHook.cs
@@ -1,6 +1,5 @@
 ﻿using Game.Enums;
 using Game.Incidents;
-using Game.Terrain;
 using Sirenix.OdinInspector;
 using Sirenix.Serialization;
 using System;
@@ -33,119 +32,6 @@ namespace Game.Simulation
 		 */
 	}
 
-	/*
- * Types of Encounters:
- * - Combat
- * - Puzzle
- * - RP
- * - Curiosity?
- * - ??
- */
-
-	//Account for multiple possible solutions
-
-	/*
-	 * Encounters contains info about whos involved, (could be people or monsters),
-	 * what those things are doing, where they are doing it, and why.
-	 * If an encounter is random and not the main part of a hook, these things
-	 * are chosen and described when they are found. But I need a way of parsing some
-	 * of that information to create a description and objectives for adventure hooks,
-	 * in the case that this encounter is the climax. 
-	 */
-
-	//Encounter Location Type [Dungeon, Overworld]
-	//Encounter Biome Type [Desert vs Forest vs ANY]
-	//Encounter Type [Puzzle vs Combat etc]
-	//Encounter approx challenge level
-
-	//Encounter description for when encountered
-	//Encounter overview for when chosen as hooks main encounter
-	//Any fun extras to find while doing encounter (like treasure or historical notes etc)
-
-	//Unrelated but need to set up a system similar to IncidentCriteria for finding threats
-	//Includes both monsters and non monster threats
-
-	[CreateAssetMenu(fileName = nameof(AdventureEncounter), menuName = "ScriptableObjects/Adventures/" + nameof(AdventureEncounter), order = 1)]
-	public class AdventureEncounter : SerializedScriptableObject
-	{
-		[PropertyOrder(-10)]
-		public string encounterTitle;
-		[PropertyOrder(-9)]
-		public EncounterLocationType encounterLocationType;
-
-		[ValueDropdown("GetEncounterTypes", IsUniqueList = true, DropdownTitle = "Encounter Types"), PropertyOrder(-8)]
-		public List<EncounterType> encounterTypes;
-
-		[ValueDropdown("GetBiomeTerrainTypes", IsUniqueList = true, DropdownTitle = "Allowed Biomes"), PropertyOrder(-7)]
-		public List<BiomeTerrainType> allowedBiomes;
-
-		[ListDrawerSettings(HideAddButton = true), PropertyOrder(0)]
-		public List<IAdventureContextCriteria> contextCriterium;
-
-		[TextArea(2, 4), PropertyOrder(0)]
-		public string encounterBlurb;
-
-		[TextArea(10,15), PropertyOrder(0)]
-		public string encounterSummary;
-
-		[TextArea(15, 20), PropertyOrder(0)]
-		public string encounterApproach;
-
-		[PropertyOrder(0), ListDrawerSettings(CustomAddFunction = "AddAct")]
-		public List<AdventureEncounterAct> encounterActs;
-
-		public AdventureEncounter()
-		{
-			encounterTypes = new List<EncounterType>();
-			allowedBiomes = new List<BiomeTerrainType>();
-			encounterActs = new List<AdventureEncounterAct>();
-			contextCriterium = new List<IAdventureContextCriteria>();
-		}
-
-		private void UpdateContainerIDs()
-		{
-			for(int i = 0; i < contextCriterium.Count; i++)
-			{
-				contextCriterium[i].ContextID = "{" + i + "}";
-			}
-		}
-
-		private IEnumerable<EncounterType> GetEncounterTypes()
-		{
-			return Enum.GetValues(typeof(EncounterType)).Cast<EncounterType>();
-		}
-
-		private IEnumerable<BiomeTerrainType> GetBiomeTerrainTypes()
-		{
-			return Enum.GetValues(typeof(BiomeTerrainType)).Cast<BiomeTerrainType>();
-		}
-
-		private void AddAct()
-		{
-			encounterActs.Add(new AdventureEncounterAct());
-		}
-
-		private void RemoveCriteria(int index)
-		{
-			contextCriterium.RemoveAt(index);
-			UpdateContainerIDs();
-		}
-
-		[ButtonGroup("1"), PropertyOrder(-1)]
-		private void AddMonster()
-		{
-			contextCriterium.Add(new MonsterCriteria());
-			UpdateContainerIDs();
-		}
-
-		[ButtonGroup("1"), PropertyOrder(-1)]
-		private void AddPerson()
-		{
-			contextCriterium.Add(new AdventureContextCriteria(typeof(Person)));
-			UpdateContainerIDs();
-		}
-	}
-
 	[HideReferenceObjectPicker]
 	public class AdventureEncounterAct
 	{
@@ -176,14 +62,16 @@ namespace Game.Simulation
 
 	public interface IAdventureContextCriteria 
 	{
-		public Type ContextType { get; set; }
+		public Type ContextType { get; }
 		public string ContextID { get; set; }
+		public IIncidentContext Context { get; set; }
+		public bool IsHistorical { get; }
 	}
 
 	[Serializable, HideReferenceObjectPicker]
-	public class AdventureContextCriteria : IAdventureContextCriteria
+	public class AdventureContextCriteria<T> : IAdventureContextCriteria where T : IIncidentContext
 	{
-		public Type ContextType { get; set; }
+		public Type ContextType => typeof(T);
 
 		public string ContextID
 		{
@@ -191,20 +79,24 @@ namespace Game.Simulation
 			set { contextID = value; }
 		}
 
-		[SerializeField, ReadOnly, HorizontalGroup, PropertyOrder(-1)]
+		public IIncidentContext Context { get; set; }
+		public bool IsHistorical => historical;
+
+		[SerializeField, ReadOnly, HorizontalGroup(LabelWidth = 120), PropertyOrder(-1)]
 		private string contextTypeName;
 		[SerializeField, ReadOnly, HorizontalGroup, PropertyOrder(-1)]
 		private string contextID;
+		[SerializeField, HorizontalGroup, PropertyOrder(-1)]
+		public bool historical;
 
-		public AdventureContextCriteria(Type contextType)
+		public AdventureContextCriteria()
 		{
-			ContextType = contextType;
-			contextTypeName = contextType.Name;
+			contextTypeName = ContextType.Name;
 		}
 	}
 
 	[Serializable, HideReferenceObjectPicker]
-	public class MonsterCriteria : AdventureContextCriteria
+	public class MonsterCriteria : AdventureContextCriteria<Monster>
 	{
 		public bool isLegendary;
 		public bool isLandDwelling;
@@ -216,7 +108,7 @@ namespace Game.Simulation
 		[ValueDropdown("GetCreatureAlignments", IsUniqueList = true, DropdownTitle = "Allowed Alignments")]
 		public List<CreatureAlignment> allowedAlignments;
 
-		public MonsterCriteria() : base(typeof(Monster))
+		public MonsterCriteria() : base()
 		{
 			allowedSizes = new List<CreatureSize>();
 			allowedTypes = new List<CreatureType>();
